// src/modules/merchandising/reports/report/style-wise-emb-status-report/emb-status-report-style-data-type.ts
export type EmbStatusReportStyleDataType = {
    PONO?: string;
    STYLEID: number;
    COLORID: number;
    STYLENO?: string;
    ITEMTYPE?: string;
    COLORNAME?: string;
    QTY: number;
    DELIVERYDATE: string;
    COMPANY_NAME?: string;
    COMPANY_ADDRESS?: string;
    PRINT_CATEGORY_ID: number;
    EMB_CATEGORY_ID: number;
    WASH_CATEGORY_ID: number;
    PRINT_EMB_CATEGORY_ID: number;
    SMOCK_CATEGORY_ID: number;
    BUYER: string;
    COMPANY_PREFIX: string;
    EMB_TYPE: string;
    CATEGORY_NAME: string;
};